avcodec/h264dec: apply H.274 film grain
Because we need access to ref frames without film grain applied, we have
to add an extra AVFrame to H264Picture to avoid messing with the
original. This requires some amount of overhead to make the reference
moves work out, but it allows us to benefit from frame multithreading
for film grain application "for free".

Unfortunately, this approach requires twice as much RAM to be constantly
allocated, due to the need for an extra buffer per H264Picture. In
theory, we could get away with freeing up this memory as soon as it's no
longer needed, but trying to call ff_thread_release_buffer() from
output_frame() simply deadlocks the decoder and I haven't figured out
why. Patches welcome(tm)

Tested on all three cases of (no fg), (fg present but exported) and (fg
present and not exported), with and without threading.

Co-authored-by: James Almer <jamrial@gmail.com>
Signed-off-by: James Almer <jamrial@gmail.com>