avcodec/libdav1d: export decoder frame delay
As this is a AV_CODEC_CAP_OTHER_THREADS decoder, threading is handled by the
underlying library. In this case, the frame delay is calculated by libdav1d
based on the values from avctx->thread_count and the private max_frame_delay
option.
With this change, this value is now exported in avcxt->delay.

Signed-off-by: James Almer <jamrial@gmail.com>